French Motor Insurance Claims Dataset

A dataset containing information about French motor insurance policies and claims, commonly used for actuarial modeling and risk assessment studies.

**IMPORTANT**: This dataset is based on commit "c49cbbb37235fc49616cac8ccac32e1491cdc619" of CASdatasets There are multiple versions of this dataset circulating online, so be weary when comparing to other versions of `freMTPL2freq`

Usage

freMTPL2freq

Format

A data frame with 677,991 rows and 10 variables:

ClaimRate

Number of claims made, at an annualised rate (numeric)

VehPower

Vehicle power rating or engine horsepower category (integer)

VehAge

Age of the vehicle in years (integer)

DrivAge

Age of the driver in years (integer)

BonusMalus

Bonus-malus coefficient, a rating factor used in French insurance where lower values indicate better driving records (integer)

VehBrand

Vehicle brand/manufacturer code (factor with levels like B6, B12, etc.)

VehGas

Type of fuel used by the vehicle (factor with levels: Regular, Diesel)

Area

Area classification where the policy holder resides (factor with levels A through F)

Density

Population density of the area where the policy holder lives (integer, inhabitants per square kilometer)

Region

French administrative region where the policy holder resides (factor with regions like Rhone-Alpes, Picardie, Aquitaine, etc.)

Source

['https://github.com/dutangc/CASdatasets/blob/master/data/freMTPL2freq.rda']

Details

This dataset is frequently used in actuarial science and insurance analytics for modeling claim frequency and severity. The ClaimRate variable represents the response (i.e. number of claims, annualised) while the other variables serve as potential predictors for risk assessment.

The BonusMalus system is a characteristic feature of French motor insurance, where coefficients typically range from 50 (best) to over 100 (worst), reflecting the policyholder's claims history.

Examples

head(freMTPL2freq)
#>        Area VehPower VehAge DrivAge BonusMalus VehBrand  VehGas Density Region
#> 412456    E       10     14      60         50       B2 Regular    5539    R11
#> 117738    E        6      4      50         53       B1 Regular    4579    R93
#> 116954    C        7      2      37         50       B3 Regular     155    R24
#> 210876    E       11      1      58         50      B12  Diesel    3301    R73
#> 640365    B        8      1      29         64      B12 Regular      51    R24
#> 153388    A        7      1      28         57       B3  Diesel      37    R11
#>        ClaimRate
#> 412456         0
#> 117738         0
#> 116954         0
#> 210876         0
#> 640365         0
#> 153388         0