Compute Beta Corrections based on SHAP values

Processes SHAP values in one-hot (wide) format to create beta coefficient corrections.

This includes:

scaling shap values of continuous variables by the predictor value for that row

migrating shap values to the bias for continuous variables where the predictor value was zero

migrating shap values to the bias for categorical variables where the predictor value was reference level

Usage

beta_corrections_derive(
  shap_wide,
  wide_input_frame,
  iblm_model,
  migrate_reference_to_bias = TRUE
)

Arguments

shap_wide

Data frame containing SHAP values from XGBoost that have been converted to wide format by [shap_to_onehot()]

wide_input_frame

Wide format input data frame (one-hot encoded).

iblm_model

Object of class 'iblm'

migrate_reference_to_bias

Logical, migrate the beta corrections to the bias for reference levels? This applied to categorical vars only. It is recommended to leave this setting on TRUE

Value

A data frame with the booster model beta corrections in one-hot (wide) format

Examples

df_list <- freMTPL2freq |>
  head(10000) |>
  dplyr::mutate(ClaimRate = round(ClaimRate)) |>
  split_into_train_validate_test()

iblm_model <- train_iblm_xgb(
  df_list,
  response_var = "ClaimRate",
  family = "poisson"
)

shap <- stats::predict(
  iblm_model$booster_model,
  newdata = xgboost::xgb.DMatrix(
    data.matrix(
      dplyr::select(df_list$test, -dplyr::all_of(iblm_model$response_var))
    )
  ),
  predcontrib = TRUE
) |> data.frame()

wide_input_frame <- data_to_onehot(df_list$test, iblm_model)

shap_wide <- shap_to_onehot(shap, wide_input_frame, iblm_model)

beta_corrections <- beta_corrections_derive(shap_wide, wide_input_frame, iblm_model)

beta_corrections |> dplyr::glimpse()
#> Rows: 1,529
#> Columns: 47
#> $ bias          <dbl> -0.153347723, -0.027749969, -0.012299433, -0.041507296, …
#> $ VehPower      <dbl> -0.0025138523, -0.0002563855, 0.0028016646, -0.002569615…
#> $ VehAge        <dbl> -0.0806419477, -0.3338305056, 0.0007959833, -0.434525936…
#> $ DrivAge       <dbl> -1.660853e-03, -1.586508e-03, 1.069323e-03, -2.835895e-0…
#> $ BonusMalus    <dbl> -1.736847e-05, -8.126060e-05, -7.244720e-05, -7.747109e-…
#> $ Density       <dbl> 1.082473e-05, -6.616692e-06, -5.356217e-04, 2.438393e-06…
#> $ AreaA         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ AreaB         <dbl> 0.0000000000, 0.0000000000, 0.0000000000, 0.0000000000, …
#> $ AreaC         <dbl> 0.0000000000, 0.0000000000, 0.0001548524, 0.0000000000, …
#> $ AreaD         <dbl> 0.0000000000, 0.0000000000, 0.0000000000, 0.0000000000, …
#> $ AreaE         <dbl> 0.0001598864, 0.0001624475, 0.0000000000, 0.0000000000, …
#> $ AreaF         <dbl> 0.0000000000, 0.0000000000, 0.0000000000, 0.0001354491, …
#> $ VehBrandB1    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ VehBrandB10   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ VehBrandB11   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ VehBrandB12   <dbl> 0.00000000, -0.01969582, 0.00000000, -0.04096698, 0.0000…
#> $ VehBrandB13   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ VehBrandB14   <dbl> 0.00000000, 0.00000000, -0.05115313, 0.00000000, 0.00000…
#> $ VehBrandB2    <dbl> 0.00000000, 0.00000000, 0.00000000, 0.00000000, -0.03359…
#> $ VehBrandB3    <dbl> 0.00000000, 0.00000000, 0.00000000, 0.00000000, 0.000000…
#> $ VehBrandB4    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ VehBrandB5    <dbl> 0.000000000, 0.000000000, 0.000000000, 0.000000000, 0.00…
#> $ VehBrandB6    <dbl> 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0…
#> $ VehGasDiesel  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ VehGasRegular <dbl> -0.041055299, 0.000000000, 0.000000000, 0.000000000, -0.…
#> $ RegionR11     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR21     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR22     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR23     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR24     <dbl> 0.00000000, 0.00000000, 0.00000000, 0.00000000, 0.000000…
#> $ RegionR25     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR26     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR31     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR41     <dbl> 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0…
#> $ RegionR42     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR43     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR52     <dbl> 0.00000000, 0.00000000, 0.02148633, 0.00000000, 0.000000…
#> $ RegionR53     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR54     <dbl> 0.00000000, 0.00000000, 0.00000000, 0.00000000, 0.019883…
#> $ RegionR72     <dbl> 0.000000000, 0.000000000, 0.000000000, 0.000000000, 0.00…
#> $ RegionR73     <dbl> 0.00000000, -0.02865498, 0.00000000, 0.00000000, 0.00000…
#> $ RegionR74     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR82     <dbl> 0.00000000, 0.00000000, 0.00000000, 0.00000000, 0.000000…
#> $ RegionR83     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ RegionR91     <dbl> 0.000000000, 0.000000000, 0.000000000, 0.000000000, 0.00…
#> $ RegionR93     <dbl> 0.05905953, 0.00000000, 0.00000000, 0.00000000, 0.000000…
#> $ RegionR94     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…