Obtain Booster Model Beta Corrections for tabular data

Creates dataframe of Shap beta corrections for each row and predictor variable of `data`

Usage

data_beta_coeff_booster(data, beta_corrections, iblm_model)

Arguments

data

A data frame containing the dataset for analysis

beta_corrections

A data frame or matrix containing beta correction values for all variables and bias

iblm_model

Object of class 'iblm'

Value

A data frame with beta coefficient corrections. The structure will be the same dimension as `data` except for a "bias" column at the start.

Examples

df_list <- freMTPL2freq |>
  head(10000) |>
  dplyr::mutate(ClaimRate = round(ClaimRate)) |>
  split_into_train_validate_test()

iblm_model <- train_iblm_xgb(
  df_list,
  response_var = "ClaimRate",
  family = "poisson"
)

explainer_outputs <- explain_iblm(iblm_model, df_list$test)

data_booster <- data_beta_coeff_booster(
  df_list$test,
  explainer_outputs$beta_corrections,
  iblm_model
)

data_booster |> dplyr::glimpse()
#> Rows: 1,576
#> Columns: 10
#> $ bias       <dbl> -0.025434082, -0.006251727, -0.022657538, -0.025341228, 0.0…
#> $ Area       <dbl> -2.554945e-04, 0.000000e+00, -7.595460e-05, 9.323163e-04, -…
#> $ VehPower   <dbl> -2.374930e-04, 4.219632e-04, -6.127541e-04, -3.928683e-04, …
#> $ VehAge     <dbl> 0.001318635, -0.043659166, 0.001579206, -0.007951507, 0.004…
#> $ DrivAge    <dbl> 1.050773e-04, -7.719393e-04, -4.158568e-04, 2.466124e-04, 1…
#> $ BonusMalus <dbl> -1.222275e-04, 1.286478e-04, -1.136246e-04, -6.451709e-05, …
#> $ VehBrand   <dbl> -0.042301517, 0.000000000, -0.042444725, -0.018660093, 0.00…
#> $ VehGas     <dbl> -0.0019267000, 0.0000000000, 0.0000000000, -0.0009197969, 0…
#> $ Density    <dbl> -1.965189e-07, -4.796051e-04, 4.612441e-05, 2.649536e-07, 7…
#> $ Region     <dbl> 0.000000e+00, -6.447542e-04, -1.102487e-04, 0.000000e+00, -…